demo-app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
anthropic>=0.39.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-multipart>=0.0.12
httpx>=0.27.0
pdfplumber>=0.11.0
python-docx>=1.1.2

# dev / test
pytest>=8.3.0
pytest-asyncio>=0.24.0
